Fleshed out bot's capabilities.
A lot of capabilities have been added, currently in testing and debugging phase. The bot now has a song queue implemented in-memory via structs, as well as it is now able to download tracks in advance to lower the time it takes to play the next track. A lot of things were made into their own function as such in a way that it makes this much easier to read than a single monstrous monolithic function.