Bug fix
Fixed a few bugs, mainly the lack of pointers to the in-memory struct which made it lose the info needed to successfully fetch tracks. Also fixed a few index out of bound errors, as well as a bug where I used os.File instead of os.Create to create a file.
As I typed this, cinnamon ran into another index out of bound error, will be fixed after this one.